I've just committed a fix that resolves critical errors in your LangChain agent and improves the code structure. Here’s a summary of what I did:
*   **Fixed the Agent Execution:** Your agent was failing because it used an outdated implementation (`ZERO_SHOT_REACT_DESCRIPTION`) that was incompatible with modern chat models and LangChain versions. I resolved this by switching to the recommended `"tool-calling"` agent type, updating the agent creation function, and replacing the deprecated `.run()` method with `.invoke()`.

*   **Refactored for Better Testability:** To make your code cleaner and easier to test, I extracted the core, non-UI functions (`get_llm`, `create_agent`) from your main Streamlit script (`super_agent-v2.py`) and moved them into a new utility file, `agent_utils.py`. This separation of concerns makes the code more modular.

*   **Incorporated Previous Fixes:** This work also builds upon my previous changes to fix an `IndentationError` and revamp the `README.md`.
